replace LoaderService with NodeService
