+++ test/unit/specs/services/node.spec.js
@@ -1,0 +1,31 @@
+import NodeService from '@/services/node'
+import store from '@/store'
+
+describe('Node Service', () => {
+  beforeAll(() => {
+    store.dispatch('network/setServer', 'https://dexplorer.ark.io:8443/api/v2')
+  })
+
+  it('should return network settings', async () => {
+    const data = await NodeService.config()
+    expect(Object.keys(data).sort()).toEqual([
+      'nethash',
+      'token',
+      'symbol',
+      'explorer',
+      'version',
+      'ports',
+      'constants',
+      'feeStatistics'
+    ].sort())
+  })
+
+  it('should return the node status', async () => {
+    const data = await NodeService.status()
+    expect(Object.keys(data).sort()).toEqual([
+      'synced',
+      'now',
+      'blocksCount'
+    ].sort())
+  })
+})
